requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pywebview>=4.0.0
cryptography>=41.0.0
pyinstaller>=6.0.0
